+++ ChangeLog.md
@@ -1,5 +1,11 @@
-## 1.6.x
+## 1.6.0b1
 
+*   Add -h and --help text to the command line tool.
+*   When portpicker is used as a command line tool from a script, allow for
+    ports chosen *without a portserver* to be kept bound to a socket by a child
+    process for a user specified timeout.  This hack *attempts* to minimize race
+    conditions. It at least prevents multiple subsequent portserver CLI
+    invocations from choosing the same port.
 *   Some pylint based refactorings to portpicker and portpicker\_test.
 *   Drop 3.6 from our CI test matrix and metadata. It probably still works
     there, but expect our unittests to include 3.7-ism's in the future. We'll